+++ server/src/services/users/paymentMethods.ts
@@ -1,4 +1,4 @@
-import { IUser } from "src/models/User";
+import { IPaymentMethod, IUser } from "src/models/User";
 import { ITransaction } from "src/models/Transaction";
 import { HydratedDocument, Types } from "mongoose";
 import { findDocumentById } from "src/utils/document";
@@ -50,5 +50,45 @@ export const sort = async (userRecord: HydratedDocument<IUser>) => {
     else pmCard.push(pm);
   }
   userRecord.paymentMethods = new Types.DocumentArray([...pmCard, ...pmAsset]);
+  await userRecord.save();
+};
+export const updatePaymentMethods = async (
+  userRecord: HydratedDocument<IUser>,
+  newPaymentMethods: any[]
+) => {
+  if (!userRecord.paymentMethods)
+    userRecord.paymentMethods = new Types.DocumentArray([]);
+
+  const pmDict: { [key: string]: IPaymentMethod } = Object.fromEntries(
+    userRecord.paymentMethods.map((pm: any) => [pm._id, pm.toObject()])
+  );
+
+  const _paymentMethods: Types.DocumentArray<IPaymentMethod> =
+    new Types.DocumentArray([]);
+
+  for (let _pm of newPaymentMethods) {
+    if (!("_id" in _pm) || !("isChecked" in _pm)) {
+      continue;
+    }
+    const key = _pm._id;
+    const isChecked = _pm.isChecked;
+
+    /* update pm */
+    const exPM = pmDict[key];
+    if (exPM) {
+      _paymentMethods.push({
+        ...exPM,
+        isChecked,
+      });
+      delete pmDict[key];
+    }
+  }
+
+  /* block removing pm */
+  if (Object.keys(pmDict).length > 0) {
+    return;
+  }
+
+  userRecord.paymentMethods = _paymentMethods;
   await userRecord.save();
 };
